requirements.txt
requests
pymongo
python-dotenv
matplotlib
opencv-python
